requirements.txt
opencv_contrib_python==4.7.0.72
opencv_python==4.6.0.66
Pillow==9.4.0
Pillow==10.4.0
Requests==2.32.3
transformers==4.44.2
